# backend/src/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cinemate API",
    "description": "Full-stack movie streaming and social recommendation platform. Catalog data from PostgreSQL; rich movie detail, cast, and trailers fetched live from TMDB API.",
    "version": "2.0.0"
  },
  "servers": [
    { "url": "/api", "description": "Reverse-proxied through Nginx" }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "HttpOnly JWT cookie set on login/register. Access token TTL 15 min; refresh token 7 days."
      }
    },
    "schemas": {
      "Movie": {
        "type": "object",
        "properties": {
          "tmdb_id":      { "type": "integer" },
          "title":        { "type": "string" },
          "year":         { "type": "integer" },
          "runtime":      { "type": "integer", "description": "Minutes" },
          "poster_path":  { "type": "string" },
          "backdrop_path":{ "type": "string" },
          "overview":     { "type": "string", "description": "Fetched live from TMDB" },
          "tagline":      { "type": "string", "description": "Fetched live from TMDB" },
          "vote_average": { "type": "number" },
          "vote_count":   { "type": "integer" },
          "avg_rating":   { "type": "number", "description": "Community average score (1–10)" },
          "rating_count": { "type": "integer" },
          "popularity":   { "type": "number" },
          "jellyfin_id":  { "type": "string", "nullable": true },
          "inDb":         { "type": "boolean", "description": "True if movie is in local catalog; false = TMDB-only" },
          "genres":       { "type": "array", "items": { "type": "string" } },
          "keywords":     { "type": "array", "items": { "type": "string" }, "description": "From TMDB" },
          "status":       { "type": "string" },
          "original_title":    { "type": "string" },
          "original_language": { "type": "string" },
          "budget":   { "type": "integer" },
          "revenue":  { "type": "integer" },
          "imdb_id":  { "type": "string" },
          "production_companies": {
            "type": "array",
            "items": { "type": "object", "properties": { "id": { "type": "integer" }, "name": { "type": "string" }, "logo_path": { "type": "string" } } }
          },
          "countries": {
            "type": "array",
            "items": { "type": "object", "properties": { "iso_code": { "type": "string" }, "name": { "type": "string" } } }
          }
        }
      },
      "CastMember": {
        "type": "object",
        "properties": {
          "id":           { "type": "integer" },
          "name":         { "type": "string" },
          "character":    { "type": "string" },
          "profile_path": { "type": "string" },
          "cast_order":   { "type": "integer" }
        }
      },
      "CrewMember": {
        "type": "object",
        "properties": {
          "id":           { "type": "integer" },
          "name":         { "type": "string" },
          "job":          { "type": "string" },
          "department":   { "type": "string" },
          "profile_path": { "type": "string" }
        }
      },
...
    }
  },
  "paths": {
...
    "/movies": {
      "get": {
        "tags": ["Movies"],
        "summary": "Browse catalog movies",
        "parameters": [
          { "name": "search",  "in": "query", "schema": { "type": "string" }, "description": "Title full-text search (trigram index)" },
          { "name": "genre",   "in": "query", "schema": { "type": "string" }, "description": "Genre name filter" },
          { "name": "year",    "in": "query", "schema": { "type": "integer" } },
          { "name": "sort",    "in": "query", "schema": { "type": "string", "enum": ["popularity","rating","newest","oldest","title","vote_count"] } },
          { "name": "page",    "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit",   "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": { "description": "Paginated movies", "content": { "application/json": { "schema": { "type": "object", "properties": { "movies": { "type": "array", "items": { "$ref": "#/components/schemas/Movie" } }, "total": { "type": "integer" }, "page": { "type": "integer" }, "pages": { "type": "integer" } } } } } }
        }
      }
    },
...
    "/movies/{tmdbId}": {
      "get": {
        "tags": ["Movies"],
        "summary": "Full movie detail — DB data + live TMDB (cast, crew, trailers, reviews, keywords). Falls back to TMDB-only if not in catalog (inDb: false).",
        "parameters": [{ "name": "tmdbId", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "Movie detail with cast, crew, comments, trailers", "content": { "application/json": { "schema": { "type": "object", "properties": { "movie": { "$ref": "#/components/schemas/Movie" }, "cast": { "type": "array", "items": { "$ref": "#/components/schemas/CastMember" } }, "crew": { "type": "array", "items": { "$ref": "#/components/schemas/CrewMember" } }, "tmdbReviews": { "type": "array", "items": { "type": "object" } }, "comments": { "type": "array", "items": { "$ref": "#/components/schemas/Comment" } }, "trailers": { "type": "array", "items": { "$ref": "#/components/schemas/Trailer" } } } } } } },
          "404": { "description": "Movie not found on TMDB" }
        }
      }
    },
...
    "/movies/{tmdbId}/subtitle": {
      "post": {
        "tags": ["Movies"],
        "summary": "Upload subtitle (SRT or VTT); SRT auto-converted to VTT on server",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "name": "tmdbId", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["content","language","label"], "properties": { "content": { "type": "string" }, "language": { "type": "string", "example": "en" }, "label": { "type": "string", "example": "English SDH" } } } } } },
        "responses": { "201": { "description": "Subtitle saved" } }
      }
    },
    "/movies/{tmdbId}/subtitle/{subtitleId}": {
      "get": {
        "tags": ["Movies"],
        "summary": "Serve subtitle file as text/vtt",
        "parameters": [
          { "name": "tmdbId",     "in": "path", "required": true, "schema": { "type": "integer" } },
          { "name": "subtitleId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "VTT file content", "content": { "text/vtt": { "schema": { "type": "string" } } } } }
      }
    },
...
    "/admin/stats": {
      "get": {
        "tags": ["Admin"],
        "summary": "Platform statistics",
        "security": [{ "cookieAuth": [] }],
        "responses": { "200": { "description": "{ movies, streamable, users, comments, ratings, subtitles, watchlist_entries }" } }
      }
    },
    "/admin/movies": {
      "get": {
        "tags": ["Admin"],
        "summary": "Paginated movie list with genre tags; type filter (real/sim/ml25m/all) and linked filter",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "page",   "in": "query", "schema": { "type": "integer" } },
          { "name": "limit",  "in": "query", "schema": { "type": "integer" } },
          { "name": "linked", "in": "query", "schema": { "type": "string", "enum": ["true","false"] } }
        ],
        "responses": { "200": { "description": "Paginated movies with genre tags" } }
      }
    },
    "/admin/movies/{tmdbId}/jellyfin": {
      "put": {
        "tags": ["Admin"],
        "summary": "Link a Jellyfin item to a movie",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "name": "tmdbId", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "jellyfinId": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Linked" } }
      },
      "delete": {
        "tags": ["Admin"],
        "summary": "Unlink Jellyfin item",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "name": "tmdbId", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Unlinked" } }
      }
    },
    "/admin/movies/{tmdbId}/upload-video": {
      "post": {
        "tags": ["Admin"],
        "summary": "Upload video to /media/movies/<Title (Year)>/; triggers Jellyfin scan + auto-links (polls 30s)",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "name": "tmdbId", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "required": true, "content": { "multipart/form-data": { "schema": { "type": "object", "properties": { "video": { "type": "string", "format": "binary" } } } } } },
        "responses": { "200": { "description": "Upload complete; Jellyfin auto-link in progress" } }
      }
    },
    "/admin/movies/{tmdbId}/subtitle": {
      "post": {
        "tags": ["Admin"],
        "summary": "Admin subtitle upload (stored in user_subtitles, visible to all users)",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "name": "tmdbId", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["content","language","label"], "properties": { "content": { "type": "string" }, "language": { "type": "string" }, "label": { "type": "string" } } } } } },
        "responses": { "201": { "description": "Subtitle saved" } }
      }
    },
    "/admin/populate-credits": {
      "post": {
        "tags": ["Admin"],
        "summary": "Trigger batch TMDB credits fetch for all catalog movies (rate-limited, ~40 min for full catalog)",
        "security": [{ "cookieAuth": [] }],
        "responses": { "200": { "description": "{ message: 'started', total }" } }
      }
    },
    "/admin/jellyfin/library": {
      "get": {
        "tags": ["Admin"],
        "summary": "List Jellyfin library movies",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "name": "search", "in": "query", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Array of { id, name, year, tmdb, path, runtime }" } }
      }
    },
...
    "/admin/comments": {
      "get": {
        "tags": ["Admin"],
        "summary": "Paginated all-comments with user and movie info",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "name": "page",   "in": "query", "schema": { "type": "integer" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "movie",  "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "{ comments[], total, pages }" } }
      }
    },
    "/admin/comments/{id}": {
      "delete": {
        "tags": ["Admin"],
        "summary": "Delete a comment",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Deleted" } }
      }
    },
    "/admin/users": {
      "get": {
        "tags": ["Admin"],
        "summary": "Paginated users with activity counts; type filter: real/sim/ml25m/all",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "name": "page",   "in": "query", "schema": { "type": "integer" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "type",   "in": "query", "schema": { "type": "string", "enum": ["real","sim","ml25m","all"] } }
        ],
        "responses": { "200": { "description": "{ users[], total, pages }" } }
      }
    },
    "/admin/users/{id}/role": {
      "put": {
        "tags": ["Admin"],
        "summary": "Change user role",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "role": { "type": "string", "enum": ["user","admin"] } } } } } },
        "responses": { "200": { "description": "Role updated" } }
      }
    },
    "/admin/users/{id}": {
      "delete": {
        "tags": ["Admin"],
        "summary": "Delete user (cannot delete own account)",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Deleted" } }
      }
    },
    "/admin/subtitles": {
      "get": {
        "tags": ["Admin"],
        "summary": "All community subtitles with uploader + movie info",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "name": "page",  "in": "query", "schema": { "type": "integer" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "{ subtitles[], total, pages }" } }
      }
    },
    "/admin/subtitles/{id}": {
      "delete": {
        "tags": ["Admin"],
        "summary": "Delete a subtitle",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Deleted" } }
      }
    },
...
  }
}